Allow option to keep temporary dirs.  Use envvar http_proxy if it exists

+++ lib/capchef.rb
@@ -120,6 +120,7 @@ module Capchef
     cmd = remote_script
     cmd += ' ' + args if args
     surun cap, cmd, options, &block
-    cap.run "rm -rf #{script_dir}", options
+  ensure
+    cap.run "rm -rf #{script_dir}", options unless ENV['CAPCHEF_KEEP_TEMP']
   end
 end
